requirements.txt
asgiref==3.8.1
Django==4.2
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
pytest-django==4.9.0
sqlparse==0.5.1
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
sqlparse==0.5.1
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.3
django-allauth==65.2.0
django-crispy-forms==2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
sqlparse==0.5.1
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.3
django-allauth==65.2.0
django-crispy-forms==2.3
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
sqlparse==0.5.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
cffi==1.17.1
crispy-bootstrap5==2024.10
Django==5.1.3
django-allauth==65.2.0
django-crispy-forms==2.3
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
sqlparse==0.5.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
cffi==1.17.1
crispy-bootstrap5==2024.10
Django==5.1.3
django-allauth==65.2.0
django-crispy-forms==2.3
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
factory_boy==3.3.1
Faker==30.8.2
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
